+++ src/components/data-table/tables/history.tsx
@@ -16,15 +16,8 @@ import {
   TableRow,
 } from "@/components/ui/table"
 import { useState } from "react"
-import { Braces, Download, File, RotateCcw, Trash2 } from "lucide-react"
 import { DataTablePagination } from "../pagination"
-import { Badge } from "../../ui/badge"
-import { DataTableViewOptions } from "../col-toggle"
-import { ButtonGroup } from "@/components/ui/button-group"
-import { Button } from "@/components/ui/button"
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
-import { DataTableProps, HistoryStatus } from "@/lib/types"
-import { capitalizeText } from "@/lib/helpers"
+import { DataTableProps } from "@/lib/types"
 
 export function HistoryTable<TData, TValue>({columns,data}: DataTableProps<TData, TValue>) {
   const [sorting, setSorting] = useState<SortingState>([]);
@@ -43,38 +36,8 @@ export function HistoryTable<TData, TValue>({columns,data}: DataTableProps<TData
     }
   })
 
-  const getBadgeValue = (cellValue: HistoryStatus) =>
-    cellValue === "warning" ? "warning" : 
-    cellValue === "error" ? "destructive" : "success";
-
   return (
     <>
-    <div className="flex items-center justify-between gap-4 w-full">
-      <ButtonGroup>
-        <Button>
-          <RotateCcw/> Refresh
-        </Button>
-        <Button variant="outline">
-          <Trash2/> Clear history
-        </Button>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="outline">
-              <Download/> Export History
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent>
-            <DropdownMenuItem>
-              <File/> Export as CSV
-            </DropdownMenuItem>
-            <DropdownMenuItem>
-              <Braces /> Export as JSON
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </ButtonGroup>
-      <DataTableViewOptions table={table}/>
-    </div>
     <div className="overflow-hidden rounded-md w-full">
       <Table>
         <TableHeader>
@@ -104,9 +67,7 @@ export function HistoryTable<TData, TValue>({columns,data}: DataTableProps<TData
               >
                 {row.getVisibleCells().map((cell) => (
                   <TableCell key={cell.id}>
-                    {cell.column.id === "status" ? <Badge variant={getBadgeValue(cell.getValue() as HistoryStatus)}>
-                      {capitalizeText(cell.getValue() as string)}
-                    </Badge> : flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
                   </TableCell>
                 ))}
               </TableRow>