+++ src/components/data-table/tables/threats.tsx
@@ -21,10 +21,8 @@ import { useState } from "react"
 import { InputGroup, InputGroupAddon, InputGroupInput } from "../../ui/input-group"
 import { SearchIcon } from "lucide-react"
 import { DataTablePagination } from "../pagination"
-import { Badge } from "../../ui/badge"
 import { DataTableViewOptions } from "../col-toggle"
-import { DataTableProps, ThreatStatus } from "@/lib/types"
-import { capitalizeText } from "@/lib/helpers"
+import { DataTableProps } from "@/lib/types"
 
 export function ThreatsTable<TData, TValue>({ columns, data, searchColumn = "displayName" }: DataTableProps<TData, TValue>) {
   const [sorting, setSorting] = useState<SortingState>([]);
@@ -46,11 +44,6 @@ export function ThreatsTable<TData, TValue>({ columns, data, searchColumn = "dis
       columnVisibility
     }
   })
-
-  const getBadgeValue = (cellValue: ThreatStatus) =>
-    cellValue ==="deleted" ? "default" :
-    cellValue ==="safe" ? "secondary" :
-    cellValue === "detected" ? "destructive" : "outline"
 
   return (
     <>
@@ -98,9 +91,7 @@ export function ThreatsTable<TData, TValue>({ columns, data, searchColumn = "dis
               >
                 {row.getVisibleCells().map((cell) => (
                   <TableCell key={cell.id}>
-                    {cell.column.id === "status" ? <Badge variant={getBadgeValue(cell.getValue() as ThreatStatus)}>
-                      {capitalizeText(cell.getValue() as string)}
-                    </Badge> : flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
                   </TableCell>
                 ))}
               </TableRow>